refactor: moving modals into uiEnhancer
