+++ src/scripts/process/events/scrappingListeners.bg.js
