+++ src/scripts/process/tasks/modals/displayInstallModals.tab.js
@@ -1,6 +1,6 @@
-import {appendModalContainer} from "../../../units/page/modals/appendModalContainer.tab";
-import ModalUpdate from "../../../units/page/modals/types/ModalUpdate.tab";
-import ModalMessage from "../../../units/page/modals/types/ModalMessage.tab";
+import {appendModalContainer} from "../../../units/page/uiEnhancer/modals/appendModalContainer.tab";
+import ModalUpdate from "../../../units/page/uiEnhancer/modals/types/ModalUpdate.tab";
+import ModalMessage from "../../../units/page/uiEnhancer/modals/types/ModalMessage.tab";
 import {getStorageData} from "../../../units/utils/chromeStorage.all";
 
 export async function launchModalIntegration(domain) {
